requirements.txt
accelerate==1.12.0
addict==2.4.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
async-timeout==4.0.3
attrs==25.4.0
backoff==2.2.1
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.10.1
contourpy==1.3.2
crcmod==1.7
cryptography==46.0.3
cycler==0.12.1
dataclasses-json==0.6.7
datasets==2.21.0
Deprecated==1.3.1
dill==0.3.8
distro==1.9.0
docling==2.69.1
docling-core==2.60.0
docling-ibm-models==3.10.3
docling-parse==4.7.3
duckduckgo_search==8.1.1
easyocr==1.7.2
effdet==0.4.1
emoji==2.15.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
faiss-cpu==1.13.2
Faker==40.1.2
fastbm25==0.0.2
filelock==3.20.3
filetype==1.2.0
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2024.6.1
google-api-core==2.29.0
google-auth==2.47.0
google-cloud-vision==3.12.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
groq==0.37.1
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
Jinja2==3.1.6
jiter==0.12.0
jmespath==0.10.0
joblib==1.5.3
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.35
langchain-experimental==0.4.1
langchain-groq==0.1.9
langchain-huggingface==1.2.0
langchain-openai==1.1.7
langchain-text-splitters==0.2.2
langchain-unstructured==1.0.1
langdetect==1.0.9
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langsmith==0.1.147
latex2mathml==3.78.1
lazy_loader==0.4
loguru==0.7.3
lxml==6.0.2
Markdown==3.10.1
markdown-it-py==4.0.0
marko==2.2.2
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
mcp==1.26.0
mdurl==0.1.2
ml_dtypes==0.5.4
modelscope==1.34.0
mpire==2.10.2
mpmath==1.3.0
msoffcrypto-tool==6.0.0
multidict==6.7.0
multiprocess==0.70.16
mypy_extensions==1.1.0
networkx==3.4.2
ninja==1.13.0
nltk==3.9.2
numpy==1.26.4
olefile==0.47
omegaconf==2.3.0
onnx==1.20.1
onnxruntime==1.23.2
openai==2.16.0
opencv-python==4.13.0.90
opencv-python-headless==4.13.0.90
openpyxl==3.1.5
orjson==3.11.5
ormsgpack==1.12.1
oss2==2.19.1
packaging==24.2
pandas==2.3.3
pdf2image==1.17.0
pdfminer.six==20251230
pdfplumber==0.11.9
pi_heif==1.1.1
pikepdf==10.2.0
pillow==11.3.0
pluggy==1.6.0
polyfactory==3.2.0
primp==0.15.0
propcache==0.4.1
proto-plus==1.27.0
protobuf==6.33.4
psutil==7.2.1
pyarrow==22.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pyclipper==1.4.0
pycocotools==2.0.11
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pylatexenc==2.10
pypandoc==1.16.2
pyparsing==3.3.2
pypdf==6.6.0
pypdfium2==5.3.0
pyreadline3==3.5.4
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-iso639==2025.11.16
python-magic==0.4.27
python-multipart==0.0.21
python-oxmsg==0.0.2
python-pptx==1.0.2
pytz==2025.2
pywin32==311
PyYAML==6.0.3
rank-bm25==0.2.2
RapidFuzz==3.14.3
rapidocr==3.5.0
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
rtree==1.4.1
safetensors==0.7.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
semchunk==2.2.2
sentence-transformers==5.2.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.45
sse-starlette==3.2.0
starlette==0.52.1
sympy==1.14.0
tabulate==0.9.0
tavily-python==0.7.21
tenacity==8.5.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tiktoken==0.12.0
timm==1.0.24
tokenizers==0.22.2
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tqdm==4.67.1
transformers==4.57.6
tree-sitter==0.24.0
tree-sitter-c==0.23.4
tree-sitter-javascript==0.23.1
tree-sitter-python==0.23.6
tree-sitter-typescript==0.23.2
typer==0.19.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
unstructured==0.18.27
unstructured-client==0.42.8
unstructured.pytesseract==0.3.15
unstructured_inference==1.1.7
urllib3==2.6.3
uuid_utils==0.13.0
uvicorn==0.40.0
webencodings==0.5.1
win32_setctime==1.2.0
wrapt==2.0.1
xlrd==2.0.2
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
